requirements.txt
asgiref==3.8.1
Django==5.1.3
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
mysqlclient==2.2.6
packaging==24.2
PyJWT==2.10.1
sqlparse==0.5.2
